+++ backend/src/main/kotlin/org/n1/av2/run/terminal/inside/CommandMoveService.kt
@@ -6,8 +6,8 @@ import org.n1.av2.hacker.skill.SkillService
 import org.n1.av2.hacker.skill.SkillType
 import org.n1.av2.layer.ice.common.IceLayer
 import org.n1.av2.layer.other.os.OsLayer
-import org.n1.av2.layer.other.shutdownAccelerator.ShutdownAcceleratorLayer
-import org.n1.av2.layer.other.shutdownAccelerator.ShutdownAcceleratorService
+import org.n1.av2.layer.other.timeradjuster.TimerAdjusterLayer
+import org.n1.av2.layer.other.timeradjuster.TimerAdjusterService
 import org.n1.av2.layer.other.tripwire.TripwireLayer
 import org.n1.av2.layer.other.tripwire.TripwireLayerService
 import org.n1.av2.platform.connection.ConnectionService
@@ -35,7 +35,7 @@ class CommandMoveService(
     private val hackerStateEntityService: HackerStateEntityService,
     private val runEntityService: RunEntityService,
     private val tripwireLayerService: TripwireLayerService,
-    private val shutdownAcceleratorService: ShutdownAcceleratorService,
+    private val timerAdjusterService: TimerAdjusterService,
     private val scanService: ScanService,
     private val userTaskRunner: UserTaskRunner,
     private val connectionService: ConnectionService,
@@ -202,7 +202,7 @@ class CommandMoveService(
         node.layers.forEach { layer ->
             when (layer) {
                 is TripwireLayer -> tripwireLayerService.hackerArrivesNode(siteId, layer, nodeId, runId)
-                is ShutdownAcceleratorLayer -> shutdownAcceleratorService.hackerArrivesNode(siteId, layer, nodeId, runId)
+                is TimerAdjusterLayer -> timerAdjusterService.hackerArrivesNode(siteId, layer, nodeId, runId)
                 else -> {} // do nothing
             }
         }