+++ backend/src/main/kotlin/org/n1/av2/site/entity/enums/LayerType.kt
@@ -13,7 +13,7 @@ enum class LayerType(
     TEXT,
     KEYSTORE,
     TRIPWIRE,
-    SHUTDOWN_ACCELERATOR,
+    TIMER_ADJUSTER,
     LOCK,
     STATUS_LIGHT,
     CORE,